[Feat] #1 스플래시 화면 적용
